Simple changes to return number of elements of the storage vector (now long).

+++ Matrix/pzskylmat.h
@@ -172,7 +172,7 @@ private:
 	int  Clear();
 	void Copy (const TPZSkylMatrix<TVar> & );
 	int Size(const int column) const {return fElem[column+1]-fElem[column];}
-	static int NumElements(const TPZVec<int> &skyline);
+	static long NumElements(const TPZVec<int> &skyline);
 	static void InitializeElem(const TPZVec<int> &skyline, TPZManVector<TVar> &storage, TPZVec<TVar *> &elem);
 	/**
      @brief Computes the highest skyline of both objects
